fix: 회원가입 엔드포인트 위치 수정
