+++ backend/main.py
@@ -43,7 +43,6 @@ app.add_middleware(
 #     response.headers["Timezone"] = "Asia/Seoul"
 #     return response
 
-
 """
 API 엔드포인트 구조
 
@@ -72,67 +71,9 @@ load_dotenv()
 
 client = OpenAI(api_key=os.getenv('OPENAI_API_KEY'))
 
-class UserCreate(BaseModel):
-    id: str
-    nickname: str
-    password: str
-    profile_image: Optional[str] = None
 
-    class Config:
-        json_schema_extra = {
-            "example": {
-                "id": "test123",
-                "nickname": "테스트유저",
-                "password": "test123!@#",
-                "profile_image": None
-            }
-        }
 
-@app.post("/user/add/", 
-    response_model=UserResponse,
-    summary="새로운 사용자 추가",
-    description="""
-    새로운 사용자를 생성합니다.
-    
-    - **id**: 고유한 사용자 아이디
-    - **nickname**: 사용자 닉네임
-    - **password**: 사용자 비밀번호
-    - **profile_image**: (선택) 프로필 이미지 URL
-    """,
-    response_description="생성된 사용자 정보",
-    tags=["Users"]
-)
-def create_user(user_data: UserCreate, db: Session = Depends(get_db)):
-    # 기존 사용자 확인
-    existing_user = db.query(models.User).filter(models.User.id == user_data.id).first()
-    if existing_user:
-        raise HTTPException(
-            status_code=400,
-            detail="이미 등록된 아이디입니다."
-        )
 
-    # 비밀번호 해싱
-    hashed_password = SecurityUtils.get_password_hash(user_data.password)
-    
-    # 사용자 생성
-    db_user = models.User(
-        id=user_data.id,
-        nickname=user_data.nickname,
-        password=hashed_password,
-        profile_image=user_data.profile_image
-    )
-
-    try:
-        db.add(db_user)
-        db.commit()
-        db.refresh(db_user)
-        return db_user
-    except Exception as e:
-        db.rollback()
-        raise HTTPException(
-            status_code=500,
-            detail="사용자 생성 중 오류가 발생했습니다."
-        )
 
 class PostCreate(BaseModel):
     user_id: int
